+++ R/tls.R
@@ -50,7 +50,7 @@
 #'
 #' @examples
 #' cfg <- tls_config()
-#' ncurl("https://www.r-project.org/", timeout = 1000L, sec = cfg)
+#' ncurl("https://www.r-project.org/", timeout = 1000L, secure = cfg)
 #'
 #' @export
 #'